# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dehasoft E-Ticaret Çözümü",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/seller/category/add": {
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Kategori oluştur",
                "description": "Yeni bir kategori oluşturur, eğer parent_id gönderilmezse ana kategori gönderilirse, gönderilen parent_id'e göre eşleşen kategorinin alt kategorisi olur.",
                "operationId": "sellerCategoryAdd",
                "requestBody": {
                    "description": "Yeni bir kategori oluşturur.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Kategori adı",
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "parent_id": {
                                        "description": "Eğer parent_id gönderilmezse ana kategori gönderilirse, gönderilen parent_id'e göre eşleşen kategorinin alt kategorisi olur.",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "icon_id": {
                                        "description": "Kategori ikonu",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Elektronik kategorisi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori oluşturuldu.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori oluşturuldu."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/seller/category/update": {
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Kategori Güncelle",
                "description": "Var olan bir kategoriyi günceller.",
                "operationId": "sellerCategoryUpdate",
                "requestBody": {
                    "description": "Var olan bir kategoriyi günceller.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Kategori adı",
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "parent_id": {
                                        "description": "Eğer parent_id gönderilmezse ana kategori gönderilirse, gönderilen parent_id'e göre eşleşen kategorinin alt kategorisi olur.",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "icon_id": {
                                        "description": "Kategori ikonu",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Elektronik kategorisi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori güncellendi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori güncellendi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/seller/category/delete": {
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Kategori Sil",
                "description": "Var olan bir kategoriyi siler.",
                "operationId": "sellerCategoryDelete",
                "requestBody": {
                    "description": "Var olan kategoriyi siler.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id"
                                ],
                                "properties": {
                                    "category_id": {
                                        "description": "Kategori Id",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori silindi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategory silindi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "deha_token": {
                "type": "apiKey",
                "description": "Login işleminden dönen tokenType ve tokeni sırasıyla boşluk bırakarak giriniz.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}